On branch master Changes to be committed: 	new file:   style.module.css
"This is why CSS modules exist. With CSS modules, you can define stylesheets that are scoped to the component.
Doing so is quite easy. All we need to do is to suffix our stylesheet file with the word "module". For example "style.module.css".
We then import the stylesheet like so:

import * as styles from './style.module.css'

When assigning the classes to elements, we need to do that a little bit different than usual.
Instead of assigning a string to className, we treat the imported styles like an object, containing all the defined classes as keys.

.class1 {
  color: red;
}

To assign the class above, we would do it like this:

<p className={styles.class1}>Example Text</p>
For CSS module to work out of the box, create-react is required.

How to use styles in React: Inline styles, CSS Modules & Styled Components
https://malcoded.com/posts/react-component-style/